Refactor, upgrade webpack
