+++ lib/ComponentData.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { Resolver } from './Resolver.js'
-import { getScript, getScriptData, isClient } from './script.js'
+import { Resolver } from './Resolver'
+import { getScript, getScriptData, isClient } from './utils'
 
 /* 
   - Handles passing of data down the tree and re-hydration between server and client
