+++ lib/Resolver.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { isClient } from './script.js'
+import { isClient } from './utils'
 import Promise from 'promise-polyfill'
 
 export class Resolver extends React.PureComponent {
