+++ lib/resolve.js
@@ -6,7 +6,7 @@ export default function resolve(component, props, req, res) {
   const methodName = 'getInitialProps';
 
   if (!component.prototype || !component.prototype.isReactComponent) {
-    throw new Error('resolve() expects a valid react component');
+    throw new Error('[React Component Data] Resolve expects a valid react component');
   }
 
   const element = React.createElement(component, props);
@@ -32,11 +32,11 @@ export default function resolve(component, props, req, res) {
   // Get components that have the data method
   const withFunction = valid.filter((component) => component[methodName]);
 
-  if (!withFunction[0] || !withFunction[0][methodName]){
+  if (!withFunction[0][methodName]){
     return null;
   } 
 
   // Call the first component's data method
   // In the future we can consider fetching data for nested components
-  return Promise.resolve(withFunction[0][methodName]({ props, req, res }));
+  return Promise.resolve(withFunction[0][methodName]({props, req, res}));
 }
