+++ package.json
@@ -33,7 +33,7 @@
     "raw-loader": "^0.5.1",
     "react-hot-loader": "3.0.0-beta.5",
     "source-map-support": "^0.4.8",
-    "webpack": "2.2.0-rc.3",
+    "webpack": "^2.2.1",
     "webpack-dev-middleware": "1.8.4",
     "webpack-hot-middleware": "2.13.0",
     "webpack-node-externals": "1.5.4"
